+++ sentry-spring-boot/src/test/kotlin/io/sentry/spring/boot/SentryAutoConfigurationTest.kt
@@ -492,6 +492,23 @@ class SentryAutoConfigurationTest {
     }
 
     @Test
+    fun `creates AOP beans to support @SentryCaptureExceptionParameter`() {
+        contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj")
+            .run {
+                assertThat(it).hasSentryExceptionParameterAdviceBeans()
+            }
+    }
+
+    @Test
+    fun `does not create AOP beans to support @SentryCaptureExceptionParameter if AOP class is missing`() {
+        contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj")
+            .withClassLoader(FilteredClassLoader(ProceedingJoinPoint::class.java))
+            .run {
+                assertThat(it).doesNotHaveSentryExceptionParameterAdviceBeans()
+            }
+    }
+
+    @Test
     fun `when tracing is enabled creates AOP beans to support @SentryTransaction`() {
         contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj", "sentry.traces-sample-rate=1.0")
             .run {
@@ -954,6 +971,20 @@ class SentryAutoConfigurationTest {
         this.doesNotHaveBean("sentrySpanPointcut")
         this.doesNotHaveBean("sentrySpanAdvice")
         this.doesNotHaveBean("sentrySpanAdvisor")
+        return this
+    }
+
+    private fun <C : ApplicationContext> ApplicationContextAssert<C>.hasSentryExceptionParameterAdviceBeans(): ApplicationContextAssert<C> {
+        this.hasBean("sentryCaptureExceptionParameterPointcut")
+        this.hasBean("sentryCaptureExceptionParameterAdvice")
+        this.hasBean("sentryCaptureExceptionParameterAdvisor")
+        return this
+    }
+
+    private fun <C : ApplicationContext> ApplicationContextAssert<C>.doesNotHaveSentryExceptionParameterAdviceBeans(): ApplicationContextAssert<C> {
+        this.doesNotHaveBean("sentryCaptureExceptionParameterPointcut")
+        this.doesNotHaveBean("sentryCaptureExceptionParameterAdvice")
+        this.doesNotHaveBean("sentryCaptureExceptionParameterAdvisor")
         return this
     }